% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connect}
\alias{connect}
\title{Connect to a database}
\usage{
connect(connection_name)
}
\arguments{
\item{connection_name}{the name of the connection you wish to connect to}
}
\value{
a DBI connecction
}
\description{
Connect to a database
}
