% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{list_connections}
\alias{list_connections}
\title{List available connections}
\usage{
list_connections()
}
\value{
a list of connections from connections file
}
\description{
List available connections
}
