% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{remove_connection}
\alias{remove_connection}
\title{Remove a connection}
\usage{
remove_connection(name)
}
\arguments{
\item{name}{name of the connection you wish to remove}
}
\description{
Remove a connection
}
